Fixed doubles rounding error

+++ src/core/Aggregators.java
@@ -578,7 +578,7 @@ public final class Aggregators {
       long total = 0;
       long numOnes = 0;
       do {
-        final double x = values.nextLongValue();
+        final int x = (int) values.nextLongValue();
         total++;
         if (x == 1) numOnes++;
       } while (values.hasNextValue());
@@ -589,7 +589,7 @@ public final class Aggregators {
       long total = 0;
       long numOnes = 0;
       do {
-        final double x = values.nextDoubleValue();
+        final int x = (int) values.nextDoubleValue();
         total++;
         if (x == 1) numOnes++;
       } while (values.hasNextValue());
@@ -619,7 +619,7 @@ public final class Aggregators {
       long total = 0;
       long numTwos = 0;
       do {
-        final double x = values.nextLongValue();
+        final int x = (int) values.nextLongValue();
         total++;
         if (x == 2) numTwos++;
       } while (values.hasNextValue());
@@ -630,7 +630,7 @@ public final class Aggregators {
       long total = 0;
       long numTwos = 0;
       do {
-        final double x = values.nextDoubleValue();
+        final int x = (int) values.nextDoubleValue();
         total++;
         if (x == 1) numTwos++;
       } while (values.hasNextValue());
